requirements.txt
numpy==1.23.5
PyMySQL==1.0.3
python-dotenv==1.0.0
